requirements.txt
python-multipart
fastapi
SQLAlchemy
pydantic[dotenv]
starlette
httpx
psycopg2
aiofiles
aioshutil
uvicorn
slowapi
passlib[bcrypt]
python-jose